feat(ui): 🎴 unify device cards across cloud + throttle
Replace 4 inconsistent device card variants (DeviceTile, DeviceConnectionCard, QuickConnectPanel, DeviceListItem) with two unified components in @repo/ui:

- ✨ **DeviceConnectCard** — single card for all connect surfaces (cloud homepage, /connect, throttle /connect, throttle home). Handles 5 states: server online/offline, device connected/disconnected w/ server online/offline. Supports `showDetailsLink` prop to toggle throttle vs cloud details button.
- ✨ **DeviceManageCard** — device management card for cloud /devices page. Fixes empty disconnect handler bug from old DeviceListItem.

🔄 **DeviceConnectionList** updates:
- Replaces DeviceTile/DeviceConnectionCard with DeviceConnectCard internally
- Handles refresh ports action internally via useDejaJS (no consumer wiring needed)
- Always shows Refresh Ports button on every device list (header OR standalone row)
- Removed dead tileMode/linkMode props and unused server-stats prop chain

🔗 **Title/icon always links to device details:**
- Cloud: navigates to /devices/:id via router
- Throttle: opens cloud details page in new named tab (`deja-cloud`)
- Shows mdi-open-in-new icon on throttle (showDetailsLink=false)

📄 **Page updates:**
- Cloud Dashboard: removed QuickConnectPanel sidebar + dead refreshPorts handler
- Cloud Layout (/devices): uses DeviceManageCard + adds Refresh Ports button
- Throttle Connect: opens device details in cloud new tab (removed modal)
- Throttle HomeView: replaced QuickConnectPanel with DeviceConnectionList

🗑️  **Deleted:** DeviceTile, DeviceConnectionCard, QuickConnectPanel, DeviceListItem

📊 **Net:** +101 / -1346 lines
