+++ packages/ui/src/DeviceConnection/DeviceConnectionList.vue
@@ -2,46 +2,40 @@
 import { computed } from 'vue'
 import type { Device, Turnout, Effect } from '@repo/modules'
 import { useLayout, useTurnouts, useEfx } from '@repo/modules'
-import DeviceConnectionCard from './DeviceConnectionCard.vue'
-import DeviceTile from '../Dashboard/DeviceTile.vue'
+import { useDejaJS } from '@repo/deja'
+import DeviceConnectCard from './DeviceConnectCard.vue'
 
 interface Props {
   devices: Device[]
   availablePorts: string[]
   availableTopics?: string[]
-  linkMode: 'page' | 'modal'
   showHeader?: boolean
-  tileMode?: boolean
+  showDetailsLink?: boolean
   serverOnline?: boolean
-  serverUptime?: string
-  connectedDeviceCount?: number
-  totalDeviceCount?: number
-  commandCount?: number
 }
 
 const props = withDefaults(defineProps<Props>(), {
   availableTopics: () => [],
   showHeader: true,
-  tileMode: false,
+  showDetailsLink: true,
   serverOnline: false,
-  serverUptime: '',
-  connectedDeviceCount: 0,
-  totalDeviceCount: 0,
-  commandCount: 0,
 })
 
 const emit = defineEmits<{
   connect: [deviceId: string, serial?: string, topic?: string]
   disconnect: [deviceId: string]
-  reconnect: [deviceId: string]
   navigate: [deviceId: string]
-  refreshPorts: []
   addDevice: []
 }>()
 
 const { getLayout } = useLayout()
 const { getTurnouts } = useTurnouts()
 const { getEffects } = useEfx()
+const { sendDejaCommand } = useDejaJS()
+
+function handleRefreshPorts() {
+  sendDejaCommand({ action: 'listPorts', payload: {} })
+}
 
 const layout = getLayout()
 
@@ -102,7 +96,7 @@ function getEffectCount(deviceId: string): number {
           variant="tonal"
           size="small"
           :disabled="!serverOnline"
-          @click="emit('refreshPorts')"
+          @click="handleRefreshPorts"
         >
           <v-icon start icon="mdi-refresh" />
           Refresh Ports
@@ -119,44 +113,36 @@ function getEffectCount(deviceId: string): number {
       </div>
     </div>
 
-    <!-- Device cards / tiles -->
-    <template v-if="tileMode">
-      <DeviceTile
-        v-for="device in sortedDevices"
-        :key="device.id"
-        :device="device"
-        :available-ports="availablePorts"
-        :available-topics="availableTopics"
-        :turnout-count="getTurnoutCount(device.id)"
-        :effect-count="getEffectCount(device.id)"
-        :track-power="device.type === 'dcc-ex' ? trackPower : null"
-        :server-uptime="device.type === 'deja-server' ? serverUptime : undefined"
-        :connected-device-count="device.type === 'deja-server' ? connectedDeviceCount : undefined"
-        :total-device-count="device.type === 'deja-server' ? totalDeviceCount : undefined"
-        :command-count="device.type === 'deja-server' ? commandCount : undefined"
-        @connect="(id, serial, topic) => emit('connect', id, serial, topic)"
-        @disconnect="(id) => emit('disconnect', id)"
-        @reconnect="(id) => emit('reconnect', id)"
-        @navigate="(id) => emit('navigate', id)"
-      />
-    </template>
-    <template v-else>
-      <DeviceConnectionCard
-        v-for="device in sortedDevices"
-        :key="device.id"
-        :device="device"
-        :available-ports="availablePorts"
-        :available-topics="availableTopics"
-        :link-mode="linkMode"
-        :turnout-count="getTurnoutCount(device.id)"
-        :effect-count="getEffectCount(device.id)"
-        :track-power="device.type === 'dcc-ex' ? trackPower : null"
-        @connect="(id, serial, topic) => emit('connect', id, serial, topic)"
-        @disconnect="(id) => emit('disconnect', id)"
-        @reconnect="(id) => emit('reconnect', id)"
-        @navigate="(id) => emit('navigate', id)"
-      />
-    </template>
+    <!-- Refresh Ports (always visible when no header) -->
+    <div v-if="!showHeader" class="d-flex justify-end mb-3">
+      <v-btn
+        variant="tonal"
+        size="small"
+        :disabled="!serverOnline"
+        @click="handleRefreshPorts"
+      >
+        <v-icon start icon="mdi-refresh" />
+        Refresh Ports
+      </v-btn>
+    </div>
+
+    <!-- Device cards -->
+    <DeviceConnectCard
+      v-for="device in sortedDevices"
+      :key="device.id"
+      class="mb-2"
+      :device="device"
+      :available-ports="availablePorts"
+      :available-topics="availableTopics"
+      :server-online="serverOnline"
+      :show-details-link="showDetailsLink"
+      :turnout-count="getTurnoutCount(device.id)"
+      :effect-count="getEffectCount(device.id)"
+      :track-power="device.type === 'dcc-ex' ? trackPower : null"
+      @connect="(id, serial, topic) => emit('connect', id, serial, topic)"
+      @disconnect="(id) => emit('disconnect', id)"
+      @navigate="(id) => emit('navigate', id)"
+    />
 
     <!-- Empty state -->
     <v-card
