+++ packages/ui/src/DeviceConnection/index.ts
@@ -1,2 +1,3 @@
-export { default as DeviceConnectionCard } from './DeviceConnectionCard.vue'
+export { default as DeviceConnectCard } from './DeviceConnectCard.vue'
 export { default as DeviceConnectionList } from './DeviceConnectionList.vue'
+export { default as DeviceManageCard } from './DeviceManageCard.vue'
